# docs/openapi/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MS Auth API",
        "description": "Serviços de autenticação, registro e MFA utilizados pelos clientes da plataforma.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Gateway principal"
        }
    ],
    "paths": {
...
        "/mfa/methods": {
            "get": {
                "tags": [
                    "MFA"
                ],
                "summary": "Lista os métodos de MFA cadastrados para o usuário autenticado.",
                "operationId": "listMfaMethods",
                "responses": {
                    "200": {
                        "description": "Lista retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MfaMethodsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/mfa/send": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Solicita o envio de um código para verificação de MFA.",
                "operationId": "sendMfaCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MfaSendRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código enviado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado ou método indisponível.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mfa/verify": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Confirma um código recebido e habilita o método.",
                "operationId": "verifyMfaCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MfaVerifyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código validado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MfaVerifyResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Código inválido ou expirado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mfa/verify-link": {
            "get": {
                "tags": [
                    "MFA"
                ],
                "summary": "Confirma um código enviado por link.",
                "operationId": "verifyMfaLink",
                "parameters": [
                    {
                        "name": "method",
                        "in": "query",
                        "description": "Método que originou o envio.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "email",
                                "sms",
                                "whatsapp"
                            ]
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "description": "Destino utilizado para enviar o código.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Código de 6 dígitos recebido.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna uma página HTML com o status da validação.",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MessageResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operação executada com sucesso."
                    }
                },
                "type": "object"
            },
...
            "MfaMethod": {
                "required": [
                    "id",
                    "method",
                    "destination",
                    "verified"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "method": {
                        "type": "string",
                        "enum": [
                            "email",
                            "sms",
                            "whatsapp"
                        ]
                    },
                    "destination": {
                        "type": "string",
                        "example": "maria@example.com"
                    },
                    "verified": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "MfaMethodsResponse": {
                "required": [
                    "methods"
                ],
                "properties": {
                    "methods": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MfaMethod"
                        }
                    }
                },
                "type": "object"
            },
            "MfaSendRequest": {
                "required": [
                    "method",
                    "destination"
                ],
                "properties": {
                    "method": {
                        "type": "string",
                        "enum": [
                            "email",
                            "sms",
                            "whatsapp"
                        ]
                    },
                    "destination": {
                        "type": "string",
                        "example": "maria@example.com"
                    }
                },
                "type": "object"
            },
            "MfaVerifyRequest": {
                "required": [
                    "method",
                    "destination",
                    "code"
                ],
                "allOf": [
                    {
                        "$ref": "#/components/schemas/MfaSendRequest"
                    },
                    {
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "123456"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "MfaVerifyResponse": {
                "required": [
                    "message",
                    "method"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Método verificado com sucesso."
                    },
                    "method": {
                        "$ref": "#/components/schemas/MfaMethod"
                    }
                },
                "type": "object"
            }
        }
    },
...
}